# docs/swagger.json
{
  "swagger": "2.0",
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:9999",
      "description": "Localhost"
    }
  ],
  "info": {
    "title": "local-ckan-harvester",
    "version": "1.0.0",
    "description": "Service for harvesting and managing ckan instance data"
  },
  "paths": {
    "/process/instance/{identifier}": {
      "get": {
        "operationId": "getProcessInstance",
        "description": "Start the processing of a ckan instance",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "process completed"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/process/package/{identifier}/{id}": {
      "get": {
        "operationId": "getProcessPackage",
        "description": "Start the processing of a ckan instance's package",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/identifier"
          },
          {
            "name": "id",
            "description": "id of ckan package for url request",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "process completed"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/process/all": {
      "get": {
        "operationId": "getProcessAll",
        "description": "Start the processing of all ckan instance",
        "produces": [
          "application/json"
        ],
        "parameters": null,
        "responses": {
          "200": {
            "description": "processes initiated"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/instance/init": {
      "get": {
        "operationId": "getInstanceInit",
        "description": "Initialize a new ckan instance",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "domain",
            "description": "Domain of the new instance, domain needs to include /api/.../ everything before /action/...",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "prefix",
            "description": "Prefix used in the domain",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "description": "CKAN version either 1 and 3 are currently supported",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "filter",
            "description": "An object which is applied agains each imported object as a filter.",
            "in": "query",
            "required": false,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Init completed"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/instance/reset/{identifier}": {
      "get": {
        "operationId": "getInstanceReset",
        "description": "Reset all tables of a ckan instance",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Reset completed"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/instance/drop/{identifier}": {
      "get": {
        "operationId": "getInstanceDrop",
        "description": "Drop all tables of a ckan instance",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Drop completed"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/master/init": {
      "get": {
        "operationId": "getMasterInit",
        "description": "Inititate the ckan management tables",
        "produces": [
          "application/json"
        ],
        "parameters": null,
        "responses": {
          "200": {
            "description": "Init completed"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/master/drop": {
      "get": {
        "operationId": "getMasterDrop",
        "description": "Drop the ckan management tables",
        "produces": [
          "application/json"
        ],
        "parameters": null,
        "responses": {
          "200": {
            "description": "Drop completed"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "identifier": {
        "name": "identifier",
        "description": "prefix (string) or ID (integer) of ckan instance.",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "500": {
        "description": "error"
      }
    }
  },
  "tags": []
}